+++ lit_gpt/utils.py
@@ -13,6 +13,7 @@ import torch
 import torch.nn as nn
 import torch.utils._device
 from lightning.fabric.strategies import FSDPStrategy
+from lightning.fabric.utilities.imports import _TORCH_GREATER_EQUAL_2_2
 from lightning.fabric.utilities.load import _lazy_load as lazy_load
 from torch.serialization import normalize_storage_type
 
@@ -346,3 +347,12 @@ def estimate_flops(model: "GPT", training: bool) -> int:
     # forward + backward
     frozen_ops_per_step = 2 if training else 1
     return ops_per_step * trainable_flops + frozen_ops_per_step * frozen_flops
+
+
+def cudagraph_mark_step() -> None:
+    if not torch._dynamo.is_compiling():
+        return
+    if _TORCH_GREATER_EQUAL_2_2:
+        torch.compiler.cudagraph_mark_step_begin()
+    else:
+        torch._inductor.cudagraph_mark_step_begin()
